Splitted provider and inputCartField
Provider to mediaquery-, productList- and invoiceProvider. And inputCartField - created invoice to the list and an errorhandler to the error loop
