+++ src/components/layout.tsx
@@ -3,7 +3,9 @@ import Navbar from "./common/navbar"
 import ContentContainer from "./contentContainer"
 import Footer from "./common/footer"
 import ScrollToTop from "./common/scrollToTop"
-import ProductListProvider, { DeviceProvider, InvoiceInfoProvider } from "./context/provider"
+import ProductListProvider from "./context/productListProvider"
+import DeviceProvider from "./context/deliveryProvider"
+import InvoiceInfoProvider from "./context/invoiceProvider"
 import CartProvider from "./context/cartProvider"
 import DeliveryProvider from "./context/deliveryProvider"
