+++ src/components/pages/admin.tsx
@@ -1,5 +1,5 @@
 import  { FC, useContext } from "react"
-import { DeviceContext} from "./../context/provider";
+import { DeviceContext} from "../context/mediaQueryProvider";
 import TabPanelAdmin from "../adminParts/tabPanelAdmin";
 import SimpleAccordion from "../adminParts/accordionAdmin";
