+++ src/components/pages/product.tsx
@@ -3,7 +3,8 @@ import { useParams, Navigate, Link } from "react-router-dom";
 import { colors } from "../../data/color";
 import { products } from "../../data/products";
 import ContentTitle from "../common/contentTitle";
-import { DeviceContext, productContext } from "../context/provider";
+import { productContext } from "../context/productListProvider";
+import { DeviceContext } from "../context/mediaQueryProvider";
 import ProductsCompare from "../product/compareProduct";
 import ProductBanner from "../product/productBanner";
 import ProductIncludeCard from "../product/productIncludeCard";
