+++ src/components/product/productBanner.tsx
@@ -3,7 +3,8 @@ import { Navigate, useParams } from "react-router-dom"
 import { colors } from "../../data/color"
 import { Product, products } from "../../data/products"
 import Lines from "../common/lines"
-import { DeviceContext, DeviceContextData, productContext } from "../context/provider"
+import { productContext } from "../context/productListProvider"
+import { DeviceContext, DeviceContextData } from "../context/mediaQueryProvider"
 import CartButton from "../interaction/cartButton"
 import { sloganStyle} from "../pages/front"
